pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools>=77",
]

[project]
name = "django-ontology"
version = "1.0.0"
description = """django-ontology is a Django application for generating sematic tripples and graphs."""
readme = "README.rst"
keywords = [
  "api",
  "django",
  "rest",
  "database",
  "ontology",
  "semantic",
]
license = "GPL-3.0"
license-files = [ "LICENSE" ]
maintainers = [
  { name = "Amir Yousefi", email = "admin@groupeffect.de" },
]
authors = [
  { name = "Amir Yousefi", email = "admin@groupeffect.de" },
]
requires-python = ">=3.12"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Web Environment",
  "Framework :: Django",
  "Framework :: Django :: 6.0",
  "Intended Audience :: Developers",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Programming Language :: Python :: Implementation :: CPython",
  "Typing :: Typed",
]
dependencies = [
  "debugpy>=1.8.20",
  "Django>=6",
  "django-filter>=25",
  "djangorestframework>=3.16",
  "rdflib>=7.6",
]
urls.Changelog = "https://github.com/groupeffect/django-ontology/blob/main/CHANGELOG.rst"
urls.Repository = "https://github.com/groupeffect/django-ontology"

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["interface*", "manage*","src/groupeffect/migrations/0001_initial.py"]

[dependency-groups]
test = [
  "asgiref>=3.6",
  "coverage[toml]",
  "pytest",
  "pytest-django",
  "pytest-randomly",
]
django60 = [ "django>=6a1,<6.1; python_version>='3.12'" ]

[tool.uv]

[tool.ruff]
lint.select = [
  # flake8-bugbear
  "B",
  # flake8-comprehensions
  "C4",
  # pycodestyle
  "E",
  # Pyflakes errors
  "F",
  # isort
  "I",
  # flake8-simplify
  "SIM",
  # flake8-tidy-imports
  "TID",
  # pyupgrade
  "UP",
  # Pyflakes warnings
  "W",
]
lint.ignore = [
  # flake8-bugbear opinionated rules
  "B9",
  # line-too-long
  "E501",
  # suppressible-exception
  "SIM105",
  # if-else-block-instead-of-if-exp
  "SIM108",
]
lint.extend-safe-fixes = [
  # non-pep585-annotation
  "UP006",
]
lint.isort.required-imports = [ "from __future__ import annotations" ]

[tool.pyproject-fmt]
max_supported_python = "3.14"

[tool.pytest]
ini_options.addopts = """\
    --strict-config
    --strict-markers
    --ds=tests.settings
    """
ini_options.django_find_project = true
ini_options.xfail_strict = true

[tool.coverage]
run.branch = true
run.data_file = ".coverage/cov"
run.parallel = true
run.source = [
  "ontology",
  "tests",
]
paths.source = [
  "src",
  ".tox/**/site-packages",
]
report.show_missing = true
report.skip_covered = true
report.skip_empty = true

[tool.mypy]
enable_error_code = [
  "ignore-without-code",
  "redundant-expr",
  "truthy-bool",
]
mypy_path = "src/"
namespace_packages = false
strict = true
warn_unreachable = true
overrides = [ { module = "tests.*", allow_untyped_defs = true } ]

[tool.rstcheck]
report_level = "ERROR"
